Working libav demuxer (including some changes to base classes).
Include extradata to AVFramedQueue. It is used to pass info from demuxers to decoders.
Include "consumed" flag to Frames. In this way, filters with multiple outputs can inform
BaseFilter of which frames have actually been written to. Before calling doProcessFrame
BaseFilter sets the flag to false and filters must set it to true if they write to each frame.
HeadDemuxerLibav retrieves the extradata from the demuxer, and passes it to the output queue.
VideoDecoderLibav gets it from the queue and passes it to libav again.
The demuxer sets the consumed flag of its output frame according to whatever kind of data it
received from libavformat.