Initial code for the StreamInfo.
Filters create a StreamInfo struct and pass it to queues, which hold a const copy.
In this way, all the different constructors for each kind of queue disappear.
Also, info regarding the stream can be passed from filter to filter (like the H264
binary format).
src directory compiles, but tests do not yet. I wanted to commit it anyway because
the changeset is becoming huge.