Implement NALU framing in the demuxer.
Now, besides converting AVCC H264 to Annex B, a single NAL unit is produced in each
buffer.
Also, a new flag has been added to StreamInfo indicating if the stream is framed or not. Since
the demuxer will always perform framing, this flag is now always true (but it is good to
have it, for future proofing).